end points

+++ src/main/java/Musicboxd/model/Music.java
@@ -17,12 +17,10 @@ public class Music {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int musicID;
-
+    private long musicID;
     private String songName;
     private String artist;
     private String genre;
     private String album;
-
 
 }
